refactor: move Home components to separate folder
