+++ src/components/Home/ResultsCard.scss
