+++ src/components/Home/ResultsList.scss
