+++ src/components/Home/SearchBar.scss
